+++ Strike/EditViewController.swift
@@ -1,0 +1,36 @@
+//
+//  EditViewController.swift
+//  Strike
+//
+//  Created by Spawn on 2016/12/09.
+//  Copyright © 2016年 Spawn. All rights reserved.
+//
+
+import UIKit
+
+class EditViewController: UITableViewCell {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
